requirements.txt
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
bleach==5.0.0
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
Django==4.0.5
django-crispy-forms==1.14.0
django-easy-select2==1.5.8
django-nested-inline==0.4.5
django-simple-history==3.1.1
django-summernote==0.8.20.0
entrypoints==0.3
grpcio==1.16.1
httplib2==0.14.0
idna==2.8
ifaddr==0.1.6
IMDbPY==6.8
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
nemo-emblems==5.0.0
netaddr==0.7.19
netifaces==0.10.4
oauthlib==3.1.0
onboard==1.4.1
packaging==20.3
PAM==0.4.2
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psutil==5.5.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyparted==3.11.2
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-debian===0.1.36ubuntu1
python-magic==0.4.16
python-xapp==2.2.1
python-xlib==0.23
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-file==1.4.3
requests-unixsocket==0.2.0
SecretStorage==2.3.1
setproctitle==1.1.10
simplejson==3.16.0
six==1.16.0
soupsieve==1.9.5
sqlparse==0.4.2
systemd-python==234
tinycss2==1.0.2
tldextract==2.2.1
tzdata==2022.1
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.1.1
urllib3==1.25.8
wadllib==1.3.3
webencodings==0.5.1
xkit==0.0.0
youtube-dl==2021.4.26
